requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt==4.0.1
"passlib[bcrypt]"
python-jose
